backend/requirements.txt
alembic==1.13.1
anyio==4.12.1
click==8.3.1
fastapi==0.109.0
freezeburn==0.6.0
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
mako==1.3.10
markupsafe==3.0.3
pillow==12.1.1
pillow-heif==0.21.0
python-multipart==0.0.20
pip==24.0
pydantic==2.5.3
pydantic_core==2.14.6
python_dotenv==1.2.1
pyyaml==6.0.3
sqlalchemy==2.0.25
starlette==0.35.1
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
